Moved the check details out of poultry as it is not exclusive to it
